pyproject.toml
[project]
name = "mygraph"
version = "0.1.0"
description = "FastAPI backend application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langgraph>=0.3.5,<1.0.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "psycopg[binary,pool]>=3.2.0",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "psycopg2>=2.9.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "graph", "schema", "db", "config", "utils", "service"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]

[tool.uv.sources]

